refactor : DTO 매핑 문제 수정 및 파일경로 변경

+++ baemin/src/main/java/com/sist/baemin/store/dto/FoodListDTO.java
@@ -5,13 +5,16 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class FoodListDTO {
     private String storeName;
-    private Integer rating;
+    private BigDecimal rating;
     private Long reviewCount;
     private Integer minimumPrice;
     private Integer deliveryFee;
